Fix live pacts, split into endpoints, move unauthed interaction to helper class, hide docker pull from logs
