+++ Makefile
@@ -15,6 +15,7 @@ install: install-node install-python install-hooks
 build: build-specification build-coordinator build-validator build-proxies
 
 test: lint validate-models check-licenses test-coordinator
+	cd tests/e2e/pact && make test
 
 publish:
 	echo Publish
