+++ azure/azure-release-template.yml
@@ -71,17 +71,17 @@ extends:
             if [ "$APIGEE_ENVIRONMENT" == "int" ]
             then
               export IDP_URL="https://nhsd-apim-testing-$APIGEE_ENVIRONMENT-ns.herokuapp.com"
-              #docker pull artronics/nhsd-login-docker:latest > /dev/null 2>&1
+              docker pull artronics/nhsd-login-docker:latest > /dev/null 2>&1
               export APIGEE_ACCESS_TOKEN=$(docker run --rm artronics/nhsd-login-docker:latest "$IDP_URL")
             elif [ "$APIGEE_ENVIRONMENT" == "ref" ]
             then
               export IDP_URL="$(REF_IDP_URL)"
               export AUTH_BEARER_TOKEN="$(REF_AUTH_BEARER_TOKEN)"
-              #docker pull booshi/nhsd-login-docker:latest > /dev/null 2>&1
+              docker pull booshi/nhsd-login-docker:latest > /dev/null 2>&1
               export APIGEE_ACCESS_TOKEN=$(docker run --rm -e AUTH_BEARER_TOKEN=$AUTH_BEARER_TOKEN booshi/nhsd-login-docker:latest "$IDP_URL")
             else
               export IDP_URL="https://nhsd-apim-testing-$APIGEE_ENVIRONMENT.herokuapp.com"
-              #docker pull artronics/nhsd-login-docker:latest > /dev/null 2>&1
+              docker pull artronics/nhsd-login-docker:latest > /dev/null 2>&1
               export APIGEE_ACCESS_TOKEN=$(docker run --rm artronics/nhsd-login-docker:latest "$IDP_URL")
             fi
           fi
@@ -113,6 +113,7 @@ extends:
           make verify-pacts
 
           # Tag
+          docker pull pactfoundation/pact-cli:latest > /dev/null 2>&1
           docker run --rm -e PACT_BROKER_BASE_URL=$PACT_BROKER_URL -e PACT_BROKER_USERNAME=$PACT_BROKER_BASIC_AUTH_USERNAME -e PACT_BROKER_PASSWORD=$PACT_BROKER_BASIC_AUTH_PASSWORD pactfoundation/pact-cli:latest broker create-version-tag --pacticipant=$PACT_PROVIDER+convert+$PACT_VERSION --version=$PACT_VERSION --tag=$APIGEE_ENVIRONMENT
           docker run --rm -e PACT_BROKER_BASE_URL=$PACT_BROKER_URL -e PACT_BROKER_USERNAME=$PACT_BROKER_BASIC_AUTH_USERNAME -e PACT_BROKER_PASSWORD=$PACT_BROKER_BASIC_AUTH_PASSWORD pactfoundation/pact-cli:latest broker create-version-tag --pacticipant=$PACT_PROVIDER+prepare+$PACT_VERSION --version=$PACT_VERSION --tag=$APIGEE_ENVIRONMENT
           docker run --rm -e PACT_BROKER_BASE_URL=$PACT_BROKER_URL -e PACT_BROKER_USERNAME=$PACT_BROKER_BASIC_AUTH_USERNAME -e PACT_BROKER_PASSWORD=$PACT_BROKER_BASIC_AUTH_PASSWORD pactfoundation/pact-cli:latest broker create-version-tag --pacticipant=$PACT_PROVIDER+process+$PACT_VERSION --version=$PACT_VERSION --tag=$APIGEE_ENVIRONMENT
