+++ tests/e2e/pact/specs/live/eps.process.spec.ts
@@ -4,12 +4,13 @@ import supertest from "supertest"
 import * as TestResources from "../../resources/test-resources"
 import { Bundle } from "../../models/fhir/fhir-resources"
 import * as LosslessJson from "lossless-json"
+import { createUnauthorisedInteraction } from "./eps-auth"
 
 jestpact.pactWith(
   {
     spec: 3,
     consumer: `nhsd-apim-eps-test-client+${process.env.PACT_VERSION}`,
-    provider: `nhsd-apim-eps+${process.env.PACT_VERSION}`,
+    provider: `nhsd-apim-eps+process+${process.env.PACT_VERSION}`,
     pactfileWriteMode: "merge"
   },
   /* eslint-disable  @typescript-eslint/no-explicit-any */
@@ -19,6 +20,21 @@ jestpact.pactWith(
       return supertest(url)
     }
 
+    const authenticationTestDescription = "a request to process an unauthorised message"
+    
+    describe("endpoint authentication e2e tests", () => {
+      test(authenticationTestDescription, async () => {
+        const apiPath = "/$process-message"
+        const interaction: InteractionObject = createUnauthorisedInteraction(authenticationTestDescription, apiPath)
+        await provider.addInteraction(interaction)
+        await client()
+          .post(apiPath)
+          .set('Content-Type', 'application/fhir+json; fhirVersion=4.0')
+          .send({})
+          .expect(401)
+      })
+    })
+
     describe("process-message e2e tests", () => {
       test.each(TestResources.processCases)("should be able to process %s", async (desc: string, message: Bundle) => {
         const apiPath = "/$process-message"
@@ -26,7 +42,7 @@ jestpact.pactWith(
         const bundle = JSON.parse(bundleStr) as Bundle
 
         const interaction: InteractionObject = {
-          state: null,
+          state: "is authenticated",
           uponReceiving: `a request to process ${desc} message to Spine`,
           withRequest: {
             headers: {
